Modified model manager to update the status of activity
